% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.target.region.R
\name{create.target.region}
\alias{create.target.region}
\title{Filters Granges object by retained mRNA list to create target region}
\usage{
create.target.region(retained.mRNA.list, gff3.filtered)
}
\arguments{
\item{retained.mRNA.list}{mRNA retained by \code{filter.mRNA.GO.list} function}

\item{gff3.filtered}{Granges object to filter by \code{retained.mRNA.list}}
}
\value{
target.region in the form of a Granges object with overlapping intervals
}
\description{
Filters Granges object by retained mRNA list to create target region
}
\examples{
target.region <- create.target.region(retained.mRNA.list, gff3.filtered)
}
