% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efetch2.R
\name{efetch2}
\alias{efetch2}
\title{Retrieve Entrez database records at NCBI in a variety of formats
code from Chris Stubben's genomes2 package modified by Jean Elbers
https://github.com/cstubben/genomes2/blob/master/R/efetch.R}
\usage{
efetch2(id, db = "pubmed", rettype = "", retmode = "text",
  showURL = FALSE, destfile, ...)
}
\arguments{
\item{id}{An EntrezHistory object or vector of Ids}

\item{db}{An Entrez database, default pubmed}

\item{rettype}{Retrieval type, see note for details}

\item{retmode}{Retrieval mode, see note for details}

\item{showURL}{display URL string}

\item{destfile}{location to save downloaded file using download.file. If missing, the url is loaded into R using readLines}

\item{...}{Other key-value pairs passed to the efetch url string, e.g seq_stop}
}
\value{
Returns efetch query
}
\description{
Retrieve Entrez database records at NCBI in a variety of formats
code from Chris Stubben's genomes2 package modified by Jean Elbers
https://github.com/cstubben/genomes2/blob/master/R/efetch.R
}
\examples{
efetch.out <- efetch2(input.R.object.split[[i]], "nucleotide", "gb", "xml")
}

