% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.gff3.R
\name{filter.gff3}
\alias{filter.gff3}
\title{Filters a GRange object by desired annotation source.}
\usage{
filter.gff3(file.name, type.to.filter)
}
\arguments{
\item{file.name}{Name of GRange object to filter}

\item{type.to.filter}{Valid type such as CDS, gene, or exon to filter by}
}
\value{
Returns a filtered gff3 file
}
\description{
Filters a GRange object by desired annotation source.
}
\examples{
gff3.filtered <- filter.gff3(gff3, 'exon')
}

