% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.mRNA.GO.list.R
\name{filter.mRNA.GO.list}
\alias{filter.mRNA.GO.list}
\title{Filters mRNA-GO list by GO id list to create Retained mRNA list}
\usage{
filter.mRNA.GO.list(mRNA.GO.list, GO.id.list)
}
\arguments{
\item{mRNA.GO.list}{R Object name of mRNA-GO list to filter}

\item{GO.id.list}{R Object name of GO id list to filter by}
}
\value{
retained.mRNA.list filtered by \code{GO.id.list}
}
\description{
Filters mRNA-GO list by GO id list to create Retained mRNA list
}
\examples{
retained.mRNA.list <- filter.mRNA.GO.list(mRNA.GO.list, GO.id.list)
}
