% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.gff3.R
\name{get.gff3}
\alias{get.gff3}
\title{Downloads the desired gff3 gene annotation file.}
\usage{
get.gff3(query.results, output.file.name)
}
\arguments{
\item{query.results}{Name given to ouput of \code{search.annot.euks} function}

\item{output.file.name}{Desired name to give to output file}
}
\value{
A GRanges object of the gene annotations
}
\description{
Downloads the desired gff3 gene annotation file.
}
\examples{
gff3 <- get.gff3(query.results, "sparrow.genome.gff3.gz")
}

