% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.maker.mRNA.translated.R
\name{get.maker.mRNA.translated}
\alias{get.maker.mRNA.translated}
\title{Get translated sequences in FASTA format from MAKER mRNA accession identifier input.}
\usage{
get.maker.mRNA.translated(input.R.object, output.R.object, input.file.name,
  output.file.name)
}
\arguments{
\item{input.R.object}{Name of R object containing mRNA accessions}

\item{output.R.object}{Name of R object to contain mRNA translated sequences}

\item{input.file.name}{Name of file with translated proteins sequences}

\item{output.file.name}{Desired name of the output fasta file}
}
\value{
An R object containing mRNA translated sequences
}
\description{
Get translated sequences in FASTA format from MAKER mRNA accession identifier input.
}
\examples{
mRNA1.translated <- get.maker.mRNA.translated(mRNA1.acc,
                                              mRNA1.translated,
                                              "maker.translated.fasta",
                                              "sparrow.mRNA1.translated.fasta")
mRNA2.translated <- get.maker.mRNA.translated(mRNA2.acc,
                                              mRNA2.translated,
                                              "maker.translated.fasta",
                                              "sparrow.mRNA2.translated.fasta")
}

