% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mRNA.translated.R
\name{get.mRNA.translated}
\alias{get.mRNA.translated}
\title{Get translated sequences in FASTA format from mRNA accession identifier input.}
\usage{
get.mRNA.translated(input.R.object, output.R.object, email.address,
  output.file.name)
}
\arguments{
\item{input.R.object}{Name of R object containing mRNA accessions}

\item{output.R.object}{Name of R object to contain mRNA translated sequences}

\item{email.address}{Your email address}

\item{output.file.name}{Desired name of the output fasta file}
}
\value{
An R object containing mRNA translated sequences
}
\description{
Get translated sequences in FASTA format from mRNA accession identifier input.
}
\examples{
mRNA1.translated <- get.mRNA.translated(mRNA1.acc,
                                      mRNA1.translated,
                                     "jelber2@lsu.edu",
                                     "sparrow.mRNA1.fasta")
mRNA2.translated <- get.mRNA.translated(mRNA2.acc,
                                      mRNA2.translated,
                                     "jelber2@lsu.edu",
                                     "sparrow.mRNA2.fasta")
}
