% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ncbi.annot.euk.genomes.R
\name{get.ncbi.annot.euk.genomes}
\alias{get.ncbi.annot.euk.genomes}
\title{Parse genomes annotated by NCBI Eukaryotic Genome Annotation Pipeline.}
\usage{
get.ncbi.annot.euk.genomes()
}
\value{
Returns all eukaryotic genomes annotated by the NCBI Eukaryotic
Genome Annotation Pipeline.
}
\description{
Parse genomes annotated by NCBI Eukaryotic Genome Annotation Pipeline.
}
\examples{
euks.filtered <- get.ncbi.annot.euk.genomes()
}

