% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.mRNA.GO.list.R
\name{make.mRNA.GO.list}
\alias{make.mRNA.GO.list}
\title{Makes mRNA-GO.list}
\usage{
make.mRNA.GO.list(path.to.GOanna.output.file, output.file.name)
}
\arguments{
\item{path.to.GOanna.output.file}{Full path to GOanna output file}

\item{output.file.name}{Name to call output file}
}
\value{
mRNA GO list
}
\description{
Makes mRNA-GO.list
}
\examples{
mRNA.GO.list <- make.mRNA.GO.list("sparrow_annot.sliminput.txt",
                                  "sparrow.mRNA.GO.list.txt")
}

