% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.maker.gff3.R
\name{read.maker.gff3}
\alias{read.maker.gff3}
\title{Reads a MAKER create gff3 file into R.}
\usage{
read.maker.gff3(gff3.file)
}
\arguments{
\item{gff3.file}{MAKER created GFF3 file (can be gzipped) that you wish to read into R}
}
\value{
A GRanges object of the gene annotations
}
\description{
Reads a MAKER create gff3 file into R.
}
\examples{
gff3 <- read.maker.gff3(gff3.file)
}
