% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save.target.region.as.bed.file.R
\name{save.target.region.as.bed.file}
\alias{save.target.region.as.bed.file}
\title{Saves final.target.region as a bed file}
\usage{
save.target.region.as.bed.file(final.target.region, output.file.name)
}
\arguments{
\item{final.target.region}{}

\item{output.file.name}{}
}
\description{
Saves final.target.region as a bed file
}
\examples{
save.target.region.as.bed.file(final.target.region, "sparrow.pigmentome.bed")
}

