% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.annot.euks.R
\name{search.annot.euks}
\alias{search.annot.euks}
\title{Searches Genomes Annotated by NCBI Eukaryotic Genome Annotation Pipeline.}
\usage{
search.annot.euks(query)
}
\arguments{
\item{query}{A species' latin name to search for}
}
\value{
Returns a query result based on \code{query}
}
\description{
Searches Genomes Annotated by NCBI Eukaryotic Genome Annotation Pipeline.
}
\examples{
query.results <- search.annot.euks("Zonotrichia albicollis")
query.results <- search.annot.euks("Canis lupus")
}

