% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upload.data.iplant.R
\name{upload.data.iplant}
\alias{upload.data.iplant}
\title{Uploads translated mRNA sequences to the iPlant's Discovery Environment}
\usage{
upload.data.iplant(user.name, password, file.name.to.upload, local.file.path)
}
\arguments{
\item{user.name}{Your iPlant user name}

\item{password}{You iPlant password}

\item{file.name.to.upload}{Name of file to upload}

\item{local.file.path}{Path of \code{file.name.to.upload}}
}
\description{
Uploads translated mRNA sequences to the iPlant's Discovery Environment
}
\examples{
uploadDataToiPlant ("jelber2",
                   "Gopherus2011",
                   "sparrow.mRNA1.fasta",
                   "C:/Users/jelber2/Dropbox/LSU/Dissertation/Manuscripts/GO2OME/rGO2TR/white_throated_sparrow/")
uploadDataToiPlant ("jelber2",
                   "Gopherus2011",
                   "sparrow.mRNA2.fasta",
                   "C:/Users/jelber2/Dropbox/LSU/Dissertation/Manuscripts/GO2OME/rGO2TR/white_throated_sparrow/")
}

