% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.target.region.as.bed.file.R
\name{save.target.region.as.bed.file}
\alias{save.target.region.as.bed.file}
\title{Saves final.target.region as a bed file}
\usage{
save.target.region.as.bed.file(final.target.region, output.file.name)
}
\arguments{
\item{final.target.region}{R object to save as BED file}

\item{output.file.name}{Name of BED file}
}
\description{
Saves final.target.region as a bed file
}
\examples{
save.target.region.as.bed.file(final.target.region, "sparrow.pigmentome.bed")
}
